Refactored to do smart contract calls on load of user page
